Extract enum helpers to use for both C# and VB
